Kernel JavaScript can be embedded in the compiler
The JavaScript that elm-node-cli will need to operate can be embedded
and loaded directly from the compiler. This avoids the whole issue of
uploading kernel code to the Elm package repository.

One issue exists though. There still needs to be a way to call the
kernel code. Normally that would look like this:
```
import Elm.Kernel.Platform
batch = Elm.Kernel.Platform.batch
```

But just like kernel code this isn't allowed, the solution is to hide
that behavior in a normal function:
```
kernelCall : String -> a
batch = kernelCall "Elm.Kernel.Platform.batch"
```

The normal Elm compiler thinks this is a regular function call and it
can thus be uploaded to the package repository, meanwhile the
elm-node-cli compiler can use this to inject calls to kernel code.

This is a bit sneaky, and perhaps to disruptive of the general Elm
ecosystem to exist... This is one of the things that will have to be
discussed with the community before release.